dhcpv6: avoid signed-shift undefined behaviour in IAID derivation
dhcpv6_generate_iface_iaid() (introduced in 591ce40) builds the IAID
from the first four bytes of the MD5 of the interface name:

    iaid = hash[0] << 24;
    iaid |= hash[1] << 16;
    ...

hash[0] is uint8_t, which is promoted to (signed) int before the
shift. When hash[0] >= 0x80, the result of 'hash[0] << 24' is
0x80000000 or higher, which does not fit in 'int' (max INT_MAX =
0x7FFFFFFF). Per C11 6.5.7p4, left-shifting a non-negative signed
value past the sign bit is undefined behaviour. Optimisers are free
to assume it never happens; in practice they have miscompiled similar
constructs in other projects.

Cast each byte to uint32_t before shifting so the shift is performed
on an unsigned type and is well-defined.

Link: https://github.com/openwrt/odhcp6c/pull/160

+++ src/dhcpv6.c
@@ -553,10 +553,10 @@ static uint32_t dhcpv6_generate_iface_iaid(const char *ifname) {
 	md5_hash(ifname, strlen(ifname), &md5);
 	md5_end(hash, &md5);
 
-	iaid = hash[0] << 24;
-	iaid |= hash[1] << 16;
-	iaid |= hash[2] << 8;
-	iaid |= hash[3];
+	iaid = (uint32_t)hash[0] << 24;
+	iaid |= (uint32_t)hash[1] << 16;
+	iaid |= (uint32_t)hash[2] << 8;
+	iaid |= (uint32_t)hash[3];
 
 	return iaid;
 }
